+++ Assets/Scripts/Scenes/GameScene/Contexts/PlayerContext/BuffSystem/Abstracts/Interfaces/IPropertyChange.cs
@@ -2,7 +2,7 @@ namespace PlayerContext.BuffSystem.Abstracts.Interfaces
 {
     internal interface IPropetyChange
     {
-        public void Increase(int value);
-        public void Decrease(int value);
+        public void Increase(float value);
+        public void Decrease(float value);
     }
 }
